Moved test/v10/python/test_init.py to 'test', and added 'test/v3x/python/test_v3x_analyzer.py'.
TODO: have 'TestV3xAnalyzer::parse_string' recognized when running pytest, and implement it.
