+++ test/v10/python/test_cqasm.py
@@ -1,4 +1,4 @@
-from .test_init import TestLibQasm
+from ...test_init import TestLibQasm
 
 
 class TestQasm(TestLibQasm):
