+++ test/v10/python/test_interface.py
@@ -1,4 +1,4 @@
-from .test_init import TestLibQasm
+from ...test_init import TestLibQasm
 
 
 class TestInterface(TestLibQasm):
