+++ test/v1x/CMakeLists.txt
@@ -1,4 +1,1 @@
-target_sources(${PROJECT_NAME}_test PRIVATE
-    "${CMAKE_CURRENT_SOURCE_DIR}/parsing.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/tutorial.cpp"
-)
+add_subdirectory(cpp)
