+++ test/v1x/cpp/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(${PROJECT_NAME}_test PRIVATE
+    "${CMAKE_CURRENT_SOURCE_DIR}/parsing.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/tutorial.cpp"
+)
